refactor(carRacing) Car list와 car 생성을 class로 구분
