+++ README.md
@@ -30,14 +30,13 @@
     - [ ] 경주 횟수를 입력 받는다.
         - [ ] 음수 예외처리
         - [ ] 문자입력의 경우 예외처리
-* 자동차 리스트
-    - [x] 배열형태의 자동차 이름을 입력 받아 각각 이름의 자동차를 생성한다.
-    - [x] null 예외처리 
 * 도구
-    - [x] 문자열을 (,)로 나눈다. 
-*  자동차
+    - [x] 문자열을 (,)로 나눈다.
+* 자동차 생성
     - [x] 문자열을 입력받는다. 
         - [x] 빈 이름 문자열 예외처리
+    - [x] 입력 받은 Name으로 자동차를 생성한다. 
+*  자동차
     - [x] 숫자를 입력받아 조건을 충족하면 움직인다. 
         - [x] 4이상이면 1 이동
         - [x] 3이하는 움직이지 않는다.
@@ -45,5 +44,5 @@
     - [ ] 자동차들의 경주를 실행한다. 
     - [ ] 자동차 경주시 랜덤 0~9의 게임포인트를 부여한다.
 * 출력 
-    - [ ] 자동차를 화면에 출력한다.
+    - [x] 매 게임을 끝날때마다 자동차 경주 중간 결과를 화면에 출력한다.
     - [ ] 우승자를 출력한다.
