+++ src/main/java/carRacing/model/RacingCarGroup.java
@@ -1,0 +1,16 @@
+package carRacing.model;
+
+import java.util.List;
+
+public class RacingCarGroup {
+
+    private List<Car> racingCars;
+
+    public RacingCarGroup(List<Car> racingCars) {
+        this.racingCars = racingCars;
+    }
+
+    public List<Car> getRacingCars() {
+        return racingCars;
+    }
+}
