+++ src/test/java/carRacing/model/RacingCarFactoryTest.java
@@ -6,21 +6,21 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class RacingCarsTest {
+class RacingCarFactoryTest {
 
     @Test
     @DisplayName("자동차 옳바른 개수의 자동차생성")
     public void raceCarsTest() {
         String[] carName = {"ferrari", "benz", "bmw"};
-        RacingCars racingCars = new RacingCars(carName);
-        assertThat(racingCars.getRacingCars().size()).isEqualTo(3);
+        RacingCarFactory racingCarFactory = new RacingCarFactory(carName);
+        assertThat(racingCarFactory.getRacingCars().size()).isEqualTo(3);
     }
 
     @Test
     @DisplayName("null 입력 예외처리")
     public void emptyCarNamesException() {
         Assertions.assertThrows(NullPointerException.class, () -> {
-            new RacingCars(null);
+            new RacingCarFactory(null);
         });
     }
 }
